requirements.txt
fastapi
uvicorn[standard]
python-binance
pandas
numpy
pydantic
PyYAML